implement comment policy

+++ app/Http/Controllers/CommentController.php
@@ -6,6 +6,7 @@ use App\Models\Article;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller implements HasMiddleware
 {
@@ -32,6 +33,8 @@ class CommentController extends Controller implements HasMiddleware
 
     public function reply(Request $request, Comment $comment)
     {
+        Gate::authorize('reply', $comment);
+
         $request->validate([
             'body' => ['required', 'string', 'min:2']
         ]);
@@ -49,6 +52,8 @@ class CommentController extends Controller implements HasMiddleware
 
     public function update(Request $request, Article $article, Comment $comment)
     {
+        Gate::authorize('update', $comment);
+
         $validatedData = $request->validate([
             'body' => ['required', 'string', 'min:3']
         ]);
@@ -60,12 +65,16 @@ class CommentController extends Controller implements HasMiddleware
 
     public function destroy(Article $article, Comment $comment)
     {
+        Gate::authorize('delete', $comment);
+
         $comment->delete();
         return back();
     }
 
     public function report(Comment $comment)
     {
+        Gate::authorize('report', $comment);
+
         if (!session()->has('reported_spams')) {
             session(['reported_spams' => []]);
         }
